Add a `ToPaths` trait for things that can render as paths
This is intended to be the new, core trait used by fart draw-ables instead of
`Shape`. Often we don't need the AABB functionality that `Shape` requires, and
the only thing that is really fundamental is the ability to render as a series
of one or more paths. And that is exactly what `ToPaths` represents.