Introduce `FartThreadRng` for an RNG that is seeded via `user_const!`
This is intended to be the new default RNG for fart programs. It is seeded from
a `user_const!`, so is controllable via the `fart serve` UI, or via environment
variable. All threads have their own RNG that is seeded with the same value, for
a little bit more determinism.